[#034 feat] user-video 연결
user-video간의 연결, objectId와 schema option중 ref 사용
- Video에서는 업로더 한 명만의 정보만 필요하므로 owner(오브젝트), User에서는 업로드한 모든 비디오들(다수)를 담아야하므로 videos(배열).
- mongoose가 부여한 랜덤id를 담아야하므로 데이터의 타입이 mongoose.Schema.Types.ObjectId 임.
- ref : 레퍼런스! 어느 schema를 참고삼아 가져오는지.
- populate() : 다른 컬렉션의 문서를 참조할 수 있게함.
ㄴ 해당 id의 데이터가 불러와짐.
- user.videos(배열)에 newVideo._id를 넣어줌.(배열이므로 push) 그 후 save()
ㄴ 여기서 user.save()를 할 때마다 비밀번호가 해싱되므로 isModified()사용 -> 비밀번호가 바뀔 경우에만 해싱.
ㄴ isModified() : 주어진 경로 중 하나라도 수정되면 true, 그렇지 않으면 false.
- 영상 업로더만이 수정 및 삭제가 가능하게끔 video.owner와 req.session._id를 비교(타입이 다르므로 String() 해줌.)
- 영상 삭제시 user.videos배열에서도 삭제해야하므로 splice() 사용후 user.save()!!

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -7,7 +8,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
   }
@@ -16,25 +17,38 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit : ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findById({ _id: id });
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: Video.formatHashtags(hashtags),
   });
+
   return res.redirect(`/videos/${id}`);
 };
 
@@ -43,16 +57,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
-  console.log(file);
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -64,7 +84,20 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  const {
+    user: { _id },
+  } = req.session;
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id, {});
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
+
   return res.redirect("/");
 };
